migrate CPU-only unit tests from openai_server to unit/entrypoints/openai
