+++ test/registered/unit/entrypoints/openai/test_protocol.py
@@ -28,10 +28,9 @@ from sglang.srt.entrypoints.openai.protocol import (
     ModelList,
     UsageInfo,
 )
-from sglang.test.ci.ci_register import register_amd_ci, register_cuda_ci
+from sglang.test.ci.ci_register import register_cpu_ci
 
-register_cuda_ci(est_time=2, suite="stage-b-test-1-gpu-small")
-register_amd_ci(est_time=10, suite="stage-b-test-1-gpu-small-amd")
+register_cpu_ci(est_time=2, suite="stage-a-test-cpu")
 
 
 class TestModelCard(unittest.TestCase):
